Fixes from https://github.com/Seeed-Studio/PN532/pull/126 and https://github.com/Seeed-Studio/PN532/pull/129

+++ src/PN532_I2C.cpp
@@ -220,7 +220,7 @@ int8_t PN532_I2C::readAckFrame()
     uint16_t time = 0;
     do
     {
-        if (_wire->requestFrom(PN532_I2C_ADDRESS, sizeof(PN532_ACK) + 1))
+        if (_wire->requestFrom(PN532_I2C_ADDRESS, (int)sizeof(PN532_ACK) + 1))
         {
             if (read() & 1)
             {          // check first byte --- status
